requirements.txt
# Core Libraries
pandas
numpy
tqdm

# File Handling
openpyxl
xlrd >= 2.0.1
python-dotenv

# LangChain Ecosystem (Updated)
langchain
langchain-openai
langgraph

# LLM Providers
openai

# Frontend UI
streamlit

# Code Execution Safety
restrictedpython

# Data Analysis & ML
scikit-learn
matplotlib
seaborn
sqlalchemy
requests

# Testing & Dev Tools
pytest
ipykernel

# Optional
# nbformat  # For future notebook support
# jupyter_client  # For safe execution context if needed
